Remove empty dir on delete managed object

+++ prymatex/support/manager.py
@@ -592,7 +592,8 @@ class PMXSupportBaseManager(object):
             os.unlink(static.path)
 
         os.unlink(filePath)
-        os.rmdir(dirname)
+        if not os.listdir(dirname):
+            os.rmdir(dirname)
 
     # ------------- BUNDLE INTERFACE
     def addBundle(self, bundle):
